requirements.txt
Flask
pandas
scikit-learn
numpy
tensorflow
Werkzeug
Pillow
flask-cors
